Crate driver_common

Device driver interfaces used by ArceOS. It provides common traits and types for implementing a device driver.

You have to use this crate with the following crates for corresponding device types:

driver_block: Common traits for block storage drivers.

driver_display: Common traits and types for graphics display drivers.

driver_net: Common traits and types for network (NIC) drivers.

Enums

DevError

The error type for device operation failures.

DeviceType

All supported device types.

Traits

BaseDriverOps

Common operations that require all device drivers to implement.

Type Aliases

DevResult

A specialized Result type for device operations.